Create python file to be run in streamlit
Create python file for stock prediction to be run in streamlit. To run the same, streamlit has to be configured properly in the system. It will have a webapp in which the user have to enter generic/unique name of stock, such as for Apple, it is AAPL, for Tesla it is TSLA. The data is extracted from yahoo finance website